requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
PyMuPDF
tqdm
groq
fastapi
uvicorn
langgraph==0.2.38
langsmith==0.1.136
langchain==0.3.3
langchain_groq==0.2.0
langchain_community==0.3.2
typing
typing_extensions==4.12.2
python-jose[cryptography]
passlib[bcrypt]
pydantic
sqlalchemy
langchain-huggingface
